% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{downloadfaf}
\alias{downloadfaf}
\title{Download FAF3 data to Rdata File}
\usage{
downloadfaf(datatypes = c("OD", "2011OD", "NETWORK", "ZONES"),
  subfoldername = "", overwrite = FALSE)
}
\arguments{
\item{datatypes}{Elements of the FAF data to download, character vector containing one or more of c("OD","2011OD","NETWORK","ZONES"), defaults to all}

\item{subfoldername}{Name for a subfolder within /data to store the files in (character string, defaults to "")}

\item{overwrite}{TRUE/FALSE to indicate whether to download and overwrite a file with the same name (defaults to FALSE)}
}
\description{
Function to download part or all of the FAF3 data from the web to the /data folder.
It uses downloadsavezip to download, read it into a data frame/spatial format, save that
to an Rdata file for later use and then remove from workspace.
}
\examples{
downloadfaf3()
}
\keyword{download}
\keyword{faf}

