% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{downloadsavezip}
\alias{downloadsavezip}
\title{Download to Rdata File}
\usage{
downloadsavezip(website, zipfilename, datafilename, datafiletype, objectname,
  outfilename, subfoldername = "", overwrite = FALSE)
}
\arguments{
\item{website}{Path to website root where data to be downloaded are (character string)}

\item{zipfilename}{Name of zip file to download (character string including for example .zip)}

\item{datafilename}{Filename (without file type ending) of the file in the zip file to extrat (character string)}

\item{datafiletype}{Fietype of the file to extract (".csv" for comma seperated variable, ".shp" for shape files)}

\item{objectname}{Object name to assign to the table of data in R (character string)}

\item{outfilename}{Filename for the Rdata file that will store the data (character string)}

\item{subfoldername}{Name for a subfolder within /data to store the files in (character string, defaults to "")}

\item{overwrite}{TRUE/FALSE to indicate whether to download and overwrite a file with the same name (defaults to FALSE)}
}
\description{
Function to download a zip file containing data in csv or shp format
from the web to the /data folder, read it into a data frame/spatial format
save that to an Rdata file for later use and then remove from workspace
}
\examples{
downloadsavezip()
}
\keyword{download}

