% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{rtransdata}
\alias{rtransdata}
\alias{rtransdata-package}
\title{Download and analyze transportation data.}
\description{
A set of functions to work  with various public transportation data sources in R,
including producing tabulations, charts, and maps to show and analyze data.
This is primarily focused on US data sources, but includes functions that might be useful
for other data sources too.
}

